+++ test/unit/fuzz-tools.test.js
@@ -278,6 +278,7 @@ import * as concussionRtsV298 from '../../lib/concussion-rts-v298.js';
 import * as cosyntropinV299 from '../../lib/cosyntropin-v299.js';
 import * as avFistulaV300 from '../../lib/av-fistula-v300.js';
 import * as drSeverityV301 from '../../lib/dr-severity-v301.js';
+import * as isisV302 from '../../lib/isis-v302.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -501,6 +502,7 @@ const MODULES = {
   'cosyntropin-v299.js': cosyntropinV299,
   'av-fistula-v300.js': avFistulaV300,
   'dr-severity-v301.js': drSeverityV301,
+  'isis-v302.js': isisV302,
   'tb-testing.js': tbTesting,
 };